+++ premium/assets/js/src/redirect-notifications.js
@@ -1,4 +1,4 @@
 var redirectFunctions = require( "./redirects/functions" );
 
-window.wpseo_undo_redirect = redirectFunctions.wpseo_undo_redirect;
-window.wpseo_create_redirect = redirectFunctions.wpseo_create_redirect;
+window.wpseoUndoRedirect = redirectFunctions.wpseoUndoRedirect;
+window.wpseoCreateRedirect = redirectFunctions.wpseoCreateRedirect;