+++ premium/classes/class-upgrade-manager.php
@@ -61,6 +61,17 @@ class WPSEO_Upgrade_Manager {
 			add_action( 'wp', array( 'WPSEO_Redirect_Upgrade', 'upgrade_3_1' ), 12 );
 			add_action( 'admin_head', array( 'WPSEO_Redirect_Upgrade', 'upgrade_3_1' ), 12 );
 		}
+
+		if ( version_compare( $version_number, '4.7', '<' ) ) {
+			add_action( 'wp', array( 'WPSEO_Premium_Prominent_Words_Versioning', 'upgrade_4_7' ), 12 );
+			add_action( 'admin_head', array( 'WPSEO_Premium_Prominent_Words_Versioning', 'upgrade_4_7' ), 12 );
+		}
+
+		if ( version_compare( $version_number, '4.8', '<' ) ) {
+			add_action( 'wp', array( 'WPSEO_Premium_Prominent_Words_Versioning', 'upgrade_4_8' ), 12 );
+			add_action( 'admin_head', array( 'WPSEO_Premium_Prominent_Words_Versioning', 'upgrade_4_8' ), 12 );
+		}
+
 	}
 
 	/**