+++ premium/classes/redirect/class-redirect-page.php
@@ -35,7 +35,7 @@ class WPSEO_Redirect_Page {
 	 */
 	public function list_table_search() {
 		if ( ( $search_string = filter_input( INPUT_POST, 's' ) ) !== null ) {
-			$url = ( $search_string !== '' ) ? add_query_arg( 's', $search_string ) : remove_query_arg( 's' );
+			$url = ( $search_string !== '' ) ? add_query_arg( 's', urlencode( $search_string ) ) : remove_query_arg( 's' );
 
 			// Do the redirect.
 			wp_redirect( $url );
@@ -57,7 +57,7 @@ class WPSEO_Redirect_Page {
 			array( 'jquery', 'jquery-ui-dialog', 'wp-util', 'underscore' ),
 			WPSEO_VERSION
 		);
-		wp_localize_script( 'wp-seo-premium-admin-redirects', 'wpseo_premium_strings', WPSEO_Premium_Javascript_Strings::strings() );
+		wp_localize_script( 'wp-seo-premium-admin-redirects', 'wpseoPremiumStrings', WPSEO_Premium_Javascript_Strings::strings() );
 		wp_localize_script( 'wp-seo-premium-admin-redirects', 'wpseoSelect2Locale', substr( WPSEO_Utils::get_user_locale(), 0, 2 ) );
 
 		wp_enqueue_style( 'wpseo-premium-redirects', plugin_dir_url( WPSEO_PREMIUM_FILE ) . 'assets/css/dist/premium-redirects-' . $version . WPSEO_CSSJS_SUFFIX . '.css', array(), WPSEO_VERSION );