+++ premium/changelog.md
@@ -1,3 +1,19 @@
+### 4.8.0: May 23rd, 2017
+* Fixes a bug where there is a double trailing slash in the javascript for the social previews that was causing errors.
+* Fixes a bug on the Redirects page where spaces were stripped when using the search field.
+* Removes the prominent words version for non-supported languages.
+* Changes the string 'current url' to 'old url' in the Search Console Redirect to match the language in the Redirect Manager. 
+* Includes every change in Yoast SEO core 4.8.0, see the [core changelog](https://wordpress.org/plugins/wordpress-seo/changelog/).
+
+### 4.7.1: May 9th, 2017
+* Includes every change in Yoast SEO core 4.7.1, see the [core changelog](https://wordpress.org/plugins/wordpress-seo/changelog/).
+
+### 4.7.0: May 2nd, 2017
+* Fixes a bug where 410 redirects could not be imported from `.htaccess`.
+* Fixes a bug where the 'email support tab' did not work on the edit post page.
+* Adds the slug of a deleted post to the redirect notification.
+* Includes every change in Yoast SEO core 4.7.0, see the [core changelog](https://wordpress.org/plugins/wordpress-seo/changelog/).
+
 ### 4.6.0: April 11th, 2017
 * Adds the ability to mark a post as cornerstone content and display this in the link suggestions box.
 * Fixes a bug where the post_modified was updated whenever the prominent words were calculated.